Refactored store to use the slices pattern
